daym/requirements.txt
django 
djangorestframework 
django-cors-headers 
djangorestframework-simplejwt
bs4
pytz
aiohttp